Added file - Android side tracking thoughts
This file is meant to be a rough outline of how the android side will
handle getting the GPS coordinates and tracking mileage for the trip.

It's very incomplete, and subject to change, but I'm adding it to the
project so the team can see what I'm thinking and possibly point out any
major flaws, missing parameters, etc., or suggest improvements.